Moved package.json into lambda function folder. Cleaned up unneeded files. Worked on bundling seeding lambda function.
