--- package.json
@@ -1,30 +1,0 @@
-{
-  "name": "kb-sport-ranking-table-converter",
-  "version": "0.0.1",
-  "description": "An AWS Lambda function that converts a Kettlebell Sport Ranking Table into a relational database",
-  "main": "index.js",
-  "repository": "https://github.com/johndrew/kb-sport-ranking-table-converter.git",
-  "author": "Taylor Johndrew <johndrew.t@gmail.com>",
-  "license": "MIT",
-  "config": {
-    "s3Bucket": "s3://johndrew-lambda-uploads/kbSportRankingTableConverter.zip"
-  },
-  "scripts": {
-    "build:data": "mkdir dist & cat data/**/*.sql > dist/data.sql",
-    "build": "docker-compose build",
-    "start": "docker-compose up",
-    "stop": "docker-compose down",
-    "test:unit": "mocha --recursive test/unit",
-    "test:integration": "mocha --recursive test/integration"
-  },
-  "dependencies": {
-    "bunyan": "^1.8.12",
-    "bunyan-format": "^0.2.1",
-    "config": "^2.0.1",
-    "mysql2": "^1.6.1"
-  },
-  "devDependencies": {
-    "mocha": "^5.2.0",
-    "sequelize": "^4.38.0"
-  }
-}
