+++ server/seedRankingTableDatabaseFunction/config/default.json
