+++ server/seedRankingTableDatabaseFunction/config/dev.json
