+++ server/seedRankingTableDatabaseFunction/seedDatabase.js
@@ -1,10 +1,10 @@
-const logger = require('../../logger');
+// const logger = require('../../logger');
 
 const seedDatabase = async (client, queries) => {
-  logger.info('Connecting to database');
+  // logger.info('Connecting to database');
   client.connect();
 
-  logger.info('Creating database');
+  // logger.info('Creating database');
   await Promise.all(queries.map((query) => client.query(query)));
 };
