+++ server/seedRankingTableDatabaseFunction/test/allTests.unit.test.js
@@ -1,0 +1,3 @@
+var context = require.context('./unit', true, /.js$/);
+context.keys().forEach(context);
+module.exports = context;
